quick script to test the fp4x2 encoding on CUDA 12.8
Summary:

Given numbers A and B, NVIDIA assumes fp4x2 is encoded as BBBBAAAA, not
AAAABBBB, i.e. A is stored in the low bits and B in high bits.

This is counterintuitive, so the script in this PR proves that this is
in fact what the inrinsics are doing.

Test Plan:

```bash
[vasiliy@devgpu023.atn1 ~/local/pytorch_scripts/nvfp4_check_format (20250321_nvfp4_check_format)]$ ./run.sh
nvcc warning : Support for offline compilation for architectures prior to '<compute/sm/lto>_75' will be removed in a future release (Use-Wno-deprecated-gpu-targets to suppress warning).
Original 1.000000 -6.000000
242:-6.0|1.0
```

Reviewers:

Subscribers:

Tasks:

Tags: